--- src/pages/Explorer/constants.ts
@@ -1,5 +1,0 @@
-export const EXPLORER_PAGE_SECTION = {
-  COLLECTIONS_SECTION: 'Collections',
-  COLLECTION_SECTION: 'Collection',
-  NFT_SECTION: 'Nft',
-}
