+++ src/pages/Mint/Collections/CollectionsItem.tsx
@@ -1,9 +1,9 @@
 import styled from 'styled-components/macro'
 
 import SampleImage from '@/assets/explorer/sample/nft_2.png'
-import { EXPLORER_PAGE_SECTION } from '@/pages/Explorer/constants'
-import type { Collection } from '@/pages/Explorer/types/collection'
-import { useContractMetadata } from '@/pages/Explorer/useContractMetadata'
+import { EXPLORER_PAGE_SECTION } from '@/pages/Mint/constants'
+import type { Collection } from '@/pages/Mint/types/collection'
+import { useContractMetadata } from '@/pages/Mint/useContractMetadata'
 
 const CollectionListItemWrapper = styled.div`
   display: flex;
