+++ src/pages/Mint/gql/NFTs.ts
