+++ src/pages/Mint/gql/collections.ts
@@ -8,7 +8,6 @@ export const GET_COLLECTIONS = gql`
       address
       metadataUri
       createdAt
-      totalSupply
     }
   }
 `
