+++ src/pages/Mint/types/collection.ts
